create controller get all order and order detail and pagination.

+++ OrangeShopBE/src/main/java/com/orange/payload/response/UserPaymentMethodDTO.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 /**
@@ -19,7 +20,7 @@ public class UserPaymentMethodDTO extends BaseEntityDTO implements Serializable 
     private PaymentType paymentType;
     @Size(max = 105)
     private String provider;
-    private Integer accountNumber;
+    private BigInteger accountNumber;
     private LocalDate expiryDate;
     private Boolean isDefault;
     private Boolean status;
